Split up LearningStatistics. Rename it for what it is, an interface between canoe and Actions.
