+++ src/riot/Components/DataDownload.riot.html
@@ -10,7 +10,7 @@
             clearInMemoryExamAnswers,
             pullExamScoresIntoMemory,
             clearInMemoryExamScores,
-        } from "Actions/Exams";
+        } from "Actions/exam";
         import { closeAndDeleteDB } from "Actions/actions_idb";
         import { dispatchLoggedInEvent } from "js/Events";
         import { isGuestUser } from "js/AuthenticationUtilities";
