+++ src/riot/Lesson/CourseExam.riot.html
@@ -70,7 +70,7 @@
         import { getLessonCardIdx, getPreviousCardsUrl, getNextCardsUrl } from "js/Routing";
         import { dispatchTestAnswerEvent, ON_ANSWERED_TEST_QUESTION } from "js/Events";
         import { URLDoesntExist, ExamIsMissingAnAnswer } from "js/Errors";
-        import { saveExamAnswer, loadExamAnswer, tallyFinalScore } from "js/LearningStatistics";
+        import { saveExamAnswer, loadExamAnswer, tallyFinalScore } from "js/ExamInterface";
 
         export default {
             state: {
