+++ src/riot/Lesson/ExamResults.riot.html
@@ -35,7 +35,8 @@
     <script>
         import LessonFrame from "RiotTags/Components/LessonFrame.riot.html";
 
-        import { markExamAsComplete, saveExamScore } from "js/LearningStatistics";
+        import { markExamAsComplete } from "js/CompletionInterface";
+        import { saveExamScore } from "js/ExamInterface";
 
         export default {
             components: {
