+++ src/riot/Lesson/LessonComplete.riot.html
@@ -66,7 +66,7 @@
     <script>
         import LessonFrame from "RiotTags/Components/LessonFrame.riot.html";
         import { getACoursesLessons } from "js/WagtailPagesAPI";
-        import { countFinishedLessonsAmongSlugs } from "js/LearningStatistics";
+        import { countFinishedLessonsAmongSlugs } from "js/CompletionInterface";
 
         export default {
             components: {
