+++ src/riot/Screens/Profile.riot.html
@@ -69,7 +69,7 @@
             countCompleteLessonsInCourses,
             getLatestCompletion,
             isCourseInProgress,
-        } from "js/LearningStatistics";
+        } from "js/CompletionInterface";
 
         export default {
             state: {
