+++ src/riot/WagtailPages/CoursePage.riot.html
@@ -47,14 +47,14 @@
         import Card from "RiotTags/Components/Card.riot.html";
         import CourseExam from "RiotTags/Lesson/CourseExam.riot.html";
 
+        import { isComplete } from "Actions/completion";
+        import { isGuestUser } from "js/AuthenticationUtilities";
         import {
             countCompleteLessonsInCourses,
             isCourseInProgress,
             isExamComplete,
-        } from "js/LearningStatistics";
-        import { isComplete } from "Actions/completion";
-        import { isGuestUser } from "js/AuthenticationUtilities";
-        import { getExamHighScore } from "js/LearningStatistics";
+        } from "js/CompletionInterface";
+        import { getExamHighScore } from "js/ExamInterface";
 
         export default {
             TRANSLATIONS: {
@@ -114,12 +114,13 @@
             },
 
             examDisabled() {
-                // exams are enabled if you complete all lessons, or you are a guest user.
+                // Exams are enabled if you complete all lessons or if you are a
+                // guest user.
                 const { wagtailPage: course } = this.props;
                 const { lessons_count } = course.data;
                 const isExamUnlocked =
                     countCompleteLessonsInCourses([course]) === lessons_count - 1;
-                return !(isExamUnlocked || isGuestUser());
+                return !isExamUnlocked && !isGuestUser();
             },
 
             isExamComplete() {
