+++ src/riot/WagtailPages/HomePage.riot.html
@@ -50,7 +50,7 @@
             countCompleteLessonsInCourses,
             getLatestCompletion,
             isCourseInProgress,
-        } from "js/LearningStatistics";
+        } from "js/CompletionInterface";
         import { ON_COMPLETION_CHANGE } from "js/Events";
         import { getTagsFromPages, doPageTagsMatchSelections } from "js/SearchAndFiltering";
